chore: separate graphql-transformer and resource-mangers for future refactor
