+++ packages/amplify-provider-awscloudformation/src/__tests__/graphql-resource-manager/dynamodb-gsi-utils.test.ts
@@ -1,4 +1,4 @@
-import * as gsiUtils from '../../graphql-transformer/dynamodb-gsi-helpers';
+import * as gsiUtils from '../../graphql-resource-manager/dynamodb-gsi-helpers';
 
 import { makeTableWithGSI } from './gsi-test-helpers';
