+++ packages/amplify-provider-awscloudformation/src/__tests__/graphql-resource-manager/gsi-test-helpers.ts
