+++ packages/amplify-provider-awscloudformation/src/graphql-resource-manager/gsi-diff-helpers.ts
