+++ packages/amplify-provider-awscloudformation/src/graphql-resource-manager/index.ts
@@ -1,0 +1,1 @@
+export { GraphQLResourceManager } from './amplify-graphql-resource-manager';
