+++ packages/amplify-provider-awscloudformation/src/graphql-transformer/index.ts
@@ -1,2 +1,1 @@
-export { GraphQLResourceManager } from './amplify-graphql-resource-manager';
 export * from './user-defined-slots';
